refactor(response): split http result into smaller manageable files feat(response): new mappable response
