+++ src/http/mod.rs
@@ -2,3 +2,4 @@ pub mod extractors;
 pub mod kernel;
 pub mod middlewares;
 pub mod server;
+pub mod response;
