+++ src/http/response/mod.rs
@@ -1,0 +1,5 @@
+pub mod defs;
+pub mod mappable;
+pub mod respond;
+pub mod result;
+pub mod r#struct;
